Removes tari_comms_middleware crate
Moved `SinkMiddleware` and `ServicePipeline` into tari_comms
