+++ base_layer/p2p/Cargo.toml
@@ -15,7 +15,6 @@ test-mocks = []
 tari_broadcast_channel = { version="^0.0",  path = "../../infrastructure/broadcast_channel" }
 tari_comms = { version = "^0.0", path = "../../comms"}
 tari_comms_dht = { version = "^0.0", path = "../../comms/dht"}
-tari_comms_middleware = { version = "^0.0", path = "../../comms/middleware"}
 tari_crypto = { version = "^0.0", path = "../../infrastructure/crypto"}
 tari_pubsub = { version = "^0.0", path = "../../infrastructure/pubsub"}
 tari_service_framework = { version = "^0.0", path = "../service_framework"}
