+++ base_layer/wallet/Cargo.toml
@@ -11,7 +11,6 @@ c_integration = []
 tari_broadcast_channel = { version="^0.0",  path = "../../infrastructure/broadcast_channel" }
 tari_comms = { path = "../../comms", version = "^0.0"}
 tari_comms_dht = { path = "../../comms/dht", version = "^0.0"}
-tari_comms_middleware = { path = "../../comms/middleware", version = "^0.0"}
 tari_crypto = { path = "../../infrastructure/crypto", version = "^0.0" }
 tari_key_manager = {path = "../key_manager", version = "^0.0"}
 tari_p2p = {path = "../p2p", version = "^0.0"}
