+++ comms/Cargo.toml
@@ -34,6 +34,7 @@ serde = "1.0.90"
 serde_derive = "1.0.90"
 snow = {version="0.6.2", features=["default-resolver"], optional = true}
 tokio = {version="^0.2", features=["blocking", "tcp", "stream", "dns", "sync", "stream"]}
+tower= "0.3.1"
 tokio-util = {version="0.2.0", features=["codec"]}
 yamux = {version="0.4.0", optional = true}
 zmq = "0.9.2"
