--- comms/middleware/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "tari_comms_middleware"
-version = "0.0.5"
-authors = ["The Tari Development Community"]
-description = "Tari comms middleware modules used to add functionality to the base tari_comms library"
-repository = "https://github.com/tari-project/tari"
-homepage = "https://tari.com"
-readme = "README.md"
-license = "BSD-3-Clause"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tari_comms = { version = "^0.0", path = "../"}
-tari_crypto = { version = "^0.0", path = "../../infrastructure/crypto"}
-tari_pubsub = { version="^0.0", path="../../infrastructure/pubsub"}
-tari_utilities = { version = "^0.0", path = "../../infrastructure/tari_util"}
-tari_shutdown = { version="^0.0",  path = "../../infrastructure/shutdown" }
-
-derive-error = "0.0.4"
-futures= {version= "^0.3.1"}
-log = "0.4.8"
-serde = "1.0.90"
-serde_derive = "1.0.90"
-tokio = "^0.2"
-tower= "0.3.0"
-rand = "0.5.5"
-
-[dev-dependencies]
-tari_test_utils = {version="^0.0", path="../../infrastructure/test_utils"}
-tari_storage = {version = "^0.0", path = "../../infrastructure/storage"}
-
-futures-test = { version = "^0.3.1" }
-lmdb-zero = "0.4.4"
-tempdir = "0.3.7"
